fix: small fixes

+++ db/repositories/pg/quiz.go
@@ -23,44 +23,48 @@ func NewQuizRepo (p *pgxpool.Pool) *PgQuizRepo{
 
 func (r *PgQuizRepo) GetQuiz(ctx context.Context)([]entities.Quiz, error){
 	var quizList []entities.Quiz;
-	rows, err := r.Pool.Query(ctx, `SELECT id, name, description FROM quiz`)
-	if rows.Err() != nil {
-		return nil, err;
+	rows, err := r.Pool.Query(ctx, `SELECT id, name, description, creator_id FROM quiz`)
+	if err != nil {
+		return nil, err
 	}
+	defer rows.Close()
+
 	for rows.Next(){
 		var quiz entities.Quiz;
-		err := rows.Scan(&quiz.ID, &quiz.Name, &quiz.Description);
+		err := rows.Scan(&quiz.ID, &quiz.Name, &quiz.Description, &quiz.CreatorID);
 		if err != nil{
 			return nil, err;
 		}
 		quizList = append(quizList, quiz);
 	}
+	if rows.Err() != nil {
+		return nil, rows.Err();
+	}
 	return quizList, nil;
 }
 
-func (r *PgQuizRepo) DeleteQuiz(ctx context.Context, quizID string)(error){
-	cmdTag, err := r.Pool.Exec(ctx, `DELETE FROM quiz WHERE id = $1`, quizID);
+func (r *PgQuizRepo) DeleteQuiz(ctx context.Context, quizID string, userID int)(error){
+	cmdTag, err := r.Pool.Exec(ctx, `DELETE FROM quiz WHERE id = $1 AND creator_id = $2`, quizID, userID);
 	if err != nil{
 		return err;
 	}
-
 	if cmdTag.RowsAffected() == 0{
 		return fmt.Errorf("quiz not found");
 	}
 	return nil;
 }
 
-func (r *PgQuizRepo) CreateQuiz(ctx context.Context, quiz_name string, quiz_description string)(entities.Quiz, error){
+func (r *PgQuizRepo) CreateQuiz(ctx context.Context, quiz_name string, quiz_description string, userID int)(entities.Quiz, error){
 	var newQuiz entities.Quiz;
 
-	err := r.Pool.QueryRow(ctx, `INSERT INTO quiz (name, description) VALUES ($1, $2) RETURNING id, name, description`, quiz_name, quiz_description).Scan(&newQuiz.ID, &newQuiz.Name, &newQuiz.Description);
+	err := r.Pool.QueryRow(ctx, `INSERT INTO quiz (name, description, creator_id) VALUES ($1, $2, $3) RETURNING id, name, description, creator_id`, quiz_name, quiz_description, userID).Scan(&newQuiz.ID, &newQuiz.Name, &newQuiz.Description, &newQuiz.CreatorID);
 	if err != nil{
 		return entities.Quiz{}, err;
 	}
 	return newQuiz, nil;
 }
 
-func (r *PgQuizRepo) UpdateQuiz(ctx context.Context, quizID string, name *string, description *string)(entities.Quiz, error){
+func (r *PgQuizRepo) UpdateQuiz(ctx context.Context, quizID string, name *string, description *string, userID int)(entities.Quiz, error){
 	query := "UPDATE quiz SET ";
 	params := []interface{}{};
 	paramCounter := 1;
@@ -86,13 +90,13 @@ func (r *PgQuizRepo) UpdateQuiz(ctx context.Context, quizID string, name *string
 		return entities.Quiz{}, fmt.Errorf("no fields to update");
 	}
 
-	query += fmt.Sprintf(" WHERE id = $%d RETURNING id, name, description", paramCounter);
-	params = append(params, quizID);
+	query += fmt.Sprintf(" WHERE id = $%d AND creator_id = $%d RETURNING id, name, description, creator_id", paramCounter, paramCounter+1);
+	params = append(params, quizID, userID);
 
 	var updatedQuiz entities.Quiz;
-	err := r.Pool.QueryRow(ctx, query, params...).Scan(&updatedQuiz.ID, &updatedQuiz.Name, &updatedQuiz.Description);
+	err := r.Pool.QueryRow(ctx, query, params...).Scan(&updatedQuiz.ID, &updatedQuiz.Name, &updatedQuiz.Description, &updatedQuiz.CreatorID);
 	if err == pgx.ErrNoRows{
-		return entities.Quiz{}, fmt.Errorf("quiz width id %s not found", quizID);
+		return entities.Quiz{}, fmt.Errorf("quiz with id %s not found", quizID);
 	}
 	if err != nil{
 		return entities.Quiz{}, err;
